backend/pyproject.toml
[project]
name = "unfold-backend"
version = "0.1.0"
description = "Neo4j graph visualizer backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "neomodel>=5.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
